chore: Rename person escort record controllers
This is a chore to retain the history of the file.
